refactor(raft): share committed-slice read between metadata_records and fetch dispatch
The controller method and the API_KEY_METADATA_FETCH server dispatch both
derived the high watermark / log-start / read-range / encode sequence
independently. Extract read_committed_slice so both call one path.
